Update `Go` code
